+++ CHANGELOG.md
@@ -11,3 +11,6 @@ Contains important changes as part of the project development adhering to [seman
 - Semantics 1 - [#7](https://github.com/SkymanOne/folidity/pull/7)
 - Parser - [#4](https://github.com/SkymanOne/folidity/pull/4)
 - Grammar specification - [#1](https://github.com/SkymanOne/folidity/pull/1)
+
+### Fixed
+- Fix Function calls - [#12](https://github.com/SkymanOne/folidity/pull/12)